Set default wavpack storage type to int

+++ include/blahdio/audio_data_format.h
@@ -22,7 +22,7 @@ struct AudioDataFormat
 			NormalizedFloat,
 		};
 
-		StorageType storage_type { StorageType::Float };
+		StorageType storage_type { StorageType::Int };
 	} wavpack;
 };
